+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/_unused/ethernet.py
@@ -2,7 +2,7 @@ from ..menus import Menus
 from ._connection_base import ConnectionMenuPage
 from .attr.margins import FIRST_LINE_Y, INFO_PAGE_MARGIN_X, SECOND_LINE_Y, THIRD_LINE_Y
 from .connection.methods import EthernetConnection
-from .helpers import draw_text
+from .render.helpers import draw_text
 
 
 class EthernetMenuPage(ConnectionMenuPage):
