+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/open_browser.py
@@ -10,7 +10,7 @@ from pitop.common.sys_info import (
 from ..menus import Menus
 from ._title_base import TitleMenuPage
 from .attr.margins import FIRST_LINE_Y, SECOND_LINE_Y, THIRD_LINE_Y
-from .helpers import draw_text
+from .render.helpers import draw_text
 
 
 class OpenBrowserMenuPage(TitleMenuPage):
