+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/welcome.py
@@ -1,7 +1,7 @@
 from ..menus import Menus
 from ._title_base import TitleMenuPage
 from .attr.margins import FIRST_LINE_Y, SECOND_LINE_Y, THIRD_LINE_Y
-from .helpers import draw_text
+from .render.helpers import draw_text
 
 
 class WelcomeMenuPage(TitleMenuPage):
